Updated Safe Mechanics
Combination has to be entered in the correct order. Dial rotates
correctly with 26 new numbers. Code has been updated for when target
numbers have been entered, but not in correct order. 1-26 can be clicked
and there is a GUI to display what number the player is on. Sounds play
for when the door opens, the dial rotates, the correct number is logged
and the wrong number is logged.